requirements.txt
gevent==1.3.7
gevent-websocket==0.10.1
python-socketio==2.0.0
bottle==0.12.13
bottle-beaker==0.1.3
bottle-cork==0.12.0
redis==3.0.1
PyMySql==0.9.2
SQLAlchemy==1.2.14
requests==2.20.1
elasticsearch-dsl==6.2.1
boto3==1.9.47
pandas==0.23.4